Create header for the interface for uring_reader I want to get to
It currently has too many responsibilities I think.
And usually not reading into start of buffer might be bad for
performance in some cases I think.

This means I have to copy and allocate for the key,
which requires setting not extreme default limit.
That shouldn't be a big problem though.

Or maybe it is:
I only need to copy when straddling a read,
but then I do need to copy everything of the line before the end.
Only need to copy when the available bytes aren't enough to get a key.
So therefore:
Need to allocate a single buffer (not per file) for unfinished line
when switching buffers, but only need to save up to key_end-1;
Othewise we don't switch until that line is selected.